Reafactor activate integration tests
